CONTRIBUTING: Update content
GitHub generates a table of content[0] for markdown documents using the
headings so there's no need to maintain it manually anymore.

There is a new Matrix chat room[1] for getting in touch with the Toolbx
developers. Advertise it instead of the alternatives.

We have new gifs from jimmac, let's use them!

https://github.com/containers/toolbox/pull/939

[0] https://github.blog/changelog/2021-04-13-table-of-contents-support-in-markdown-files
[1] https://matrix.to/#/#toolbx:matrix.org