cmd/initContainer: Include execution error in message
This will pair with a future change to `shell.Run()` so that we capture
the child process stderr.

But actually this change on its own is enough since `shell.Run()`
provides an error message when the invoked command was not found or when
some other unknown error has happened.

Before:
  Error: failed to remove password for user walters`
After:
  Error: failed to remove password for user walters: passwd(1) not found`

which helps immediately pinpoint the problem.

I didn't try to go through and change *all* the `shell.Run()`
invocations, but if accepted I may do it (or someone else can).

https://github.com/containers/toolbox/pull/945