profile.d: Add "istoolbx" alias
Adds a short command alias that can be run both from the host as well as
the toolbx container. If it is executed from the host, no output is
printed and an error code 1 is returned. If run from a toolbx, prints
out the name of the current toolbx and returns 0 instead.

Signed-off-by: Andreas Hartmann <andreas.h@rtmann.de>