feat: allow custom hostnames for new containers
The new `--hostname` flag is now available when invoking `toolbox
create` in order to create a container with a custom hostname. The new
flag does not enforce network-resolvability of any custom hostnames,
but it does require that the hostname is valid, per RFC 1123.

To preserve backward-compatibility, the hostname will still default to
"toolbox" if the flag is not specified.

Signed-off-by: Buckley Ross <buckleyross42@gmail.com>
Change-Id: Ib3c0706e3a93a5748453e39998b9893e3e57c305