src: Assign tests to test suites
The project already has a number of different tests plugged into the
build system. To have greater control over which tests are executed,
assign them into different suites. 'linting' for formatting and linting
and 'unit' for unit tests and generally self-contained tests testing
a small portion of Toolbx. Also, in a follow-up commit the system tests
will be plugged into Meson and having a way of running only some of the
tests will be needed.

https://github.com/containers/toolbox/pull/1062